Footprint and labelling
Added the Teensy40 footprint into the custom-footprints.pretty folder, which will allow us to link the teensy40 footprint to the schematic we have. will still need to be readdressed.

Changed the raw text on the schematic to labels so that they will be recognized as connections rather than just text.